+++ selfdrive/test/test_onroad.py
@@ -55,6 +55,7 @@ PROCS = {
   "selfdrive.locationd.lagd": 11.0,
   "selfdrive.ui.soundd": 3.0,
   "selfdrive.ui.feedback.feedbackd": 1.0,
+  "selfdrive.nap_dash.server": 4.0,
   "selfdrive.monitoring.dmonitoringd": 4.0,
   "system.proclogd": 7.0,
   "system.logmessaged": 1.0,